:truck: Create request interfaces file
