+++ api/interfaces/requests.go
@@ -1,0 +1,12 @@
+package interfaces
+
+type SignUpForm struct {
+	Username string `json:"username"`
+	Email    string `json:"email"`
+	Password string `json:"password"`
+}
+
+type LogInForm struct {
+	Email    string `json:"email"`
+	Password string `json:"password"`
+}
